# backend/api/cart/v1/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: CartService API
    version: 0.0.1
paths:
    /v1/cart:
        get:
            tags:
                - CartService
            description: 获取购物车信息
            operationId: CartService_GetCart
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetCartResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - CartService
            description: 新增购物车商品
            operationId: CartService_UpsertItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpsertItemReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpsertItemResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - CartService
            description: 清空购物车
            operationId: CartService_EmptyCart
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EmptyCartResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/create:
        post:
            tags:
                - CartService
            operationId: CartService_CreateCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCartReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateCartResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/item/check:
        post:
            tags:
                - CartService
            operationId: CartService_CheckCartItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CheckCartItemReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckCartItemResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/item/uncheck:
        post:
            tags:
                - CartService
            operationId: CartService_UncheckCartItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UncheckCartItemReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UncheckCartItemResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/item/{product_id}:
        delete:
            tags:
                - CartService
            description: 删除购物车商品
            operationId: CartService_RemoveCartItem
            parameters:
                - name: product_id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RemoveCartItemResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/list:
        get:
            tags:
                - CartService
            operationId: CartService_ListCarts
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCartsResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/cart/order:
        post:
            tags:
                - CartService
            operationId: CartService_CreateOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateOrderReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateOrderResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Cart:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/CartItem'
            description: 购物车信息
        CartItem:
            type: object
            properties:
                product_id:
                    type: integer
                    format: uint32
                quantity:
                    type: integer
                    format: int32
                selected:
                    type: boolean
            description: 购物车商品
        CheckCartItemReq:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
                product_id:
                    type: integer
                    format: uint32
            description: 选中购物车商品
        CheckCartItemResp:
            type: object
            properties:
                success:
                    type: boolean
        CreateCartReq:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
            description: 新建购物车
        CreateCartResp:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        CreateOrderReq:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
            description: 创建订单
        CreateOrderResp:
            type: object
            properties:
                success:
                    type: boolean
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/CartItem'
        EmptyCartResp:
            type: object
            properties:
                success:
                    type: boolean
        GetCartResp:
            type: object
            properties:
                cart:
                    $ref: '#/components/schemas/Cart'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListCartsResp:
            type: object
            properties:
                carts:
                    type: array
                    items:
                        $ref: '#/components/schemas/Cart'
        RemoveCartItemResp:
            type: object
            properties:
                success:
                    type: boolean
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UncheckCartItemReq:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
                product_id:
                    type: integer
                    format: uint32
            description: 取消选中购物车商品
        UncheckCartItemResp:
            type: object
            properties:
                success:
                    type: boolean
        UpsertItemReq:
            type: object
            properties:
                owner:
                    type: string
                name:
                    type: string
                item:
                    $ref: '#/components/schemas/CartItem'
            description: 新增购物车商品
        UpsertItemResp:
            type: object
            properties:
                success:
                    type: boolean
tags:
    - name: CartService